chore: move investor-foxy into lib
