--- packages/investor-foxy/.gitignore
@@ -1,2 +1,0 @@
-coverage
-dist
