--- packages/investor-foxy/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "@shapeshiftoss/investor-foxy",
-  "version": "8.0.1",
-  "description": "An API layer for interaction with the ShapeShift FOXy contract",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "directories": {
-    "lib": "dist",
-    "src": "src"
-  },
-  "files": [
-    "dist"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "build": "yarn clean && yarn run -T tsc --build",
-    "clean": "rm -rf dist tsconfig.tsbuildinfo",
-    "dev": "yarn run -T tsc --build --watch",
-    "cli": "yarn build && yarn node dist/foxycli.js"
-  },
-  "dependencies": {
-    "@shapeshiftoss/caip": "workspace:^",
-    "@shapeshiftoss/chain-adapters": "workspace:^",
-    "@shapeshiftoss/logger": "workspace:^",
-    "@shapeshiftoss/types": "workspace:^",
-    "readline-sync": "^1.4.10"
-  },
-  "devDependencies": {
-    "@types/readline-sync": "^1.4.4"
-  }
-}
