--- packages/investor-foxy/sample.env
@@ -1,2 +1,0 @@
-MNEMONIC='match with local foxy (only used in cli)'
-ARCHIVE_NODE=https://link-to-archive-node.com
