--- packages/investor-foxy/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../../tsconfig.packages.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist"
-  },
-  "include": ["src/**/*"],
-  "exclude": ["dist"],
-  "references": [
-    { "path": "../caip" },
-    { "path": "../chain-adapters" },
-    { "path": "../logger" },
-    { "path": "../types" },
-  ]
-}
