+++ src/lib/investor/investor-foxy/abi/foxy-staking-abi.ts
