+++ src/lib/investor/investor-foxy/api/foxy-types.ts
