+++ src/lib/investor/investor-foxy/constants/allowance.ts
