+++ src/lib/investor/investor-foxy/constants/enums.ts
