+++ src/pages/Dashboard/components/DashboardHeader.tsx
@@ -1,6 +1,5 @@
 import { ArrowDownIcon, ArrowUpIcon } from '@chakra-ui/icons'
 import { Button, Container, Flex, Skeleton, Stack, useColorModeValue } from '@chakra-ui/react'
-import { bnOrZero } from '@shapeshiftoss/investor-foxy'
 import { useCallback, useEffect, useMemo, useRef } from 'react'
 import { IoSwapVerticalSharp } from 'react-icons/io5'
 import { useTranslate } from 'react-polyglot'
@@ -14,6 +13,7 @@ import { useFeatureFlag } from 'hooks/useFeatureFlag/useFeatureFlag'
 import { useModal } from 'hooks/useModal/useModal'
 import { useWallet } from 'hooks/useWallet/useWallet'
 import { makeBlockiesUrl } from 'lib/blockies/makeBlockiesUrl'
+import { bnOrZero } from 'lib/investor/investor-foxy'
 import {
   selectClaimableRewards,
   selectEarnBalancesFiatAmountFull,
