+++ src/plugins/foxPage/components/AssetActions.tsx
@@ -15,7 +15,6 @@ import {
 import type { AssetId } from '@shapeshiftoss/caip'
 import { foxAssetId } from '@shapeshiftoss/caip'
 import { supportsETH } from '@shapeshiftoss/hdwallet-core/dist/wallet'
-import { foxyAddresses } from '@shapeshiftoss/investor-foxy'
 import { DefiProvider } from 'features/defi/contexts/DefiManagerProvider/DefiCommon'
 import isEqual from 'lodash/isEqual'
 import qs from 'qs'
@@ -28,6 +27,7 @@ import { Text } from 'components/Text/Text'
 import { WalletActions } from 'context/WalletProvider/actions'
 import { useModal } from 'hooks/useModal/useModal'
 import { useWallet } from 'hooks/useWallet/useWallet'
+import { foxyAddresses } from 'lib/investor/investor-foxy'
 import { getMixPanel } from 'lib/mixpanel/mixPanelSingleton'
 import { MixPanelEvents } from 'lib/mixpanel/types'
 import { TradeCard } from 'pages/Dashboard/TradeCard'
