+++ src/state/apis/foxy/foxyApiSingleton.ts
@@ -1,8 +1,8 @@
 import type { EvmBaseAdapter } from '@shapeshiftoss/chain-adapters'
-import { foxyAddresses, FoxyApi } from '@shapeshiftoss/investor-foxy'
 import { KnownChainIds } from '@shapeshiftoss/types'
 import { getConfig } from 'config'
 import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
+import { foxyAddresses, FoxyApi } from 'lib/investor/investor-foxy'
 
 // don't export me, access me through the getter
 let _foxyApi: FoxyApi | undefined = undefined
