+++ src/state/slices/opportunitiesSlice/resolvers/yearn/index.ts
@@ -1,9 +1,9 @@
 import type { ToAssetIdArgs } from '@shapeshiftoss/caip'
 import { fromAssetId, toAssetId } from '@shapeshiftoss/caip'
-import { bnOrZero } from '@shapeshiftoss/investor-foxy'
 import { USDC_PRECISION } from 'constants/constants'
 import { DefiProvider, DefiType } from 'features/defi/contexts/DefiManagerProvider/DefiCommon'
 import { getYearnInvestor } from 'features/defi/contexts/YearnProvider/yearnInvestorSingleton'
+import { bnOrZero } from 'lib/investor/investor-foxy'
 import {
   selectAssetById,
   selectFeatureFlags,
