+++ src/state/slices/opportunitiesSlice/utils.ts
@@ -1,10 +1,10 @@
 import type { AccountId, AssetId, ChainId } from '@shapeshiftoss/caip'
 import { fromAccountId, fromAssetId, toAccountId, toAssetId } from '@shapeshiftoss/caip'
-import { bnOrZero } from '@shapeshiftoss/investor-foxy'
 import type { MarketData } from '@shapeshiftoss/types'
 import { DefiProvider, DefiType } from 'features/defi/contexts/DefiManagerProvider/DefiCommon'
 import type { Asset } from 'lib/asset-service'
 import { bn } from 'lib/bignumber/bignumber'
+import { bnOrZero } from 'lib/investor/investor-foxy'
 import { fromBaseUnit } from 'lib/math'
 
 import { foxEthAssetIds, STAKING_ID_DELIMITER } from './constants'
